finished cpp basic converter

+++ src/convertors/convert_frame.h
@@ -3,6 +3,6 @@
 
 cv::Mat convert_image_to_minecraft_blocks(const cv::Mat& img);
 
-cv::Mat downscale_image(const std::string& image_path, int scale_factor = 16);
+cv::Mat downscale_image(const std::string& image_path, size_t scale_factor = 16);
 
 #endif // CONVERT_FRAME_H
